+++ Documentation/Changelog.md
@@ -6,6 +6,8 @@
 
 Release date: not released yet
 
+- Improved API of `ExecuteOnce<T>` primitive
+- Improved API of `EvaluateOnce<T>` primitive
 - Added `GetRelativePath(string, string)` method to `IReadOnlyFileSystemView` VFS interface
 - Added `System.Environment.ProcessPath` polyfill property
 - Fixed MSB3277 warning about `System.ValueTuple` assembly that occurred for .NET Framework target frameworks (caused by `Microsoft.Bcl.Memory` package)